% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDepVector.R
\name{checkDepVector}
\alias{checkDepVector}
\title{Perform check on the true class label vector}
\usage{
checkDepVector(depTest)
}
\arguments{
\item{depTest}{Vector with true data labels (outcome values)}
}
\description{
Perform check on the true class label vector.
}
\examples{
## Load response modeling predictions
data("response")
## Apply checkDepVector checking function
checkDepVector(response$test[,1])

}
\author{
Koen W. De Bock, \email{kdebock@audencia.com}
}
